task1/app/requirements.txt
psycopg[binary]~=3.2
PyYAML~=6.0

